CALCULATION AGENT/requirements.txt
google-adk
pandas
numpy
openpyxl
streamlit
plotly
yfinance
fastapi
uvicorn
python-dotenv
